lots of committed files... but hardly anything to see ;)
Process handling/Command calls simplified for external users (see example files):
- e.g.  Proc.get(Commands.XQUERY, query)...  -> new XQuery(query).execute(context)
- same parser used for Command and Query parsing to improve future GUI code completion

+++ src/org/basex/data/Data.java
@@ -413,7 +413,7 @@ public abstract class Data  {
    * @param val value to be inserted
    * @param kind node kind
    */
-  public abstract void insert(int pre, int par, byte[] val, byte kind);
+  public abstract void insert(int pre, int par, byte[] val, int kind);
   
   /**
    * Inserts an attribute.
